+++ esp-hal-common/src/embassy/time_driver_systimer.rs
@@ -20,7 +20,7 @@ pub struct EmbassyTimer {
 
 const ALARM_STATE_NONE: AlarmState = AlarmState::new();
 
-embassy_time::time_driver_impl!(static DRIVER: EmbassyTimer = EmbassyTimer {
+embassy_time_driver::time_driver_impl!(static DRIVER: EmbassyTimer = EmbassyTimer {
     alarms: Mutex::new([ALARM_STATE_NONE; ALARM_COUNT]),
     alarm0: unsafe { Alarm::<_, 0>::conjure() },
     alarm1: unsafe { Alarm::<_, 1>::conjure() },
@@ -88,7 +88,7 @@ impl EmbassyTimer {
 
     pub(crate) fn set_alarm(
         &self,
-        alarm: embassy_time::driver::AlarmHandle,
+        alarm: embassy_time_driver::AlarmHandle,
         timestamp: u64,
     ) -> bool {
         critical_section::with(|_cs| {